fix(docs): removed colons from image filenames (#32)
> Hmm, I'm having issues as I'm on Windows. The images are causing
problems because the ':' characters in filenames aren't manageable on
Windows. I can't publish the JSON file without also publishing the fact
that I've deleted the images. :/ And if I create the registry file on my
machine, I'm concerned that the JSON will contain Windows line breaks.

_Originally posted by @Drosscend in
https://github.com/kianbazza/ui/issues/30#issuecomment-2773245459_

This PR addresses an issue for Windows contributors. The colons present
in certain filenames causes issues with Biome.